extract actor selector component
