+++ backend/src/main/resources/web/index.html
@@ -30,23 +30,23 @@
     </ul>
     <div class="tab-content">
         <div class="tab-pane fade active in panel panel-default" id="actorselection">
-            <div class="panel-body" id="actortree">
-                <table class="table table-striped table-hover ">
-                    <thead>
-                    <tr>
-                        <th></th>
-                        <th>Actor
-                            <p style="float:right;">
-                                <input id="actorfilter-regex" size="12" type="text" placeholder="Filter..."/>
-                                <a href="#" id="actorfilter-select-all">all</a> | <a href="#" id="actorfilter-select-none">none</a>
-                            </p>
-                        </th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    </tbody>
-                </table>
-            </div>
+            <!--<div class="panel-body" id="actortree">-->
+                <!--<table class="table table-striped table-hover ">-->
+                    <!--<thead>-->
+                    <!--<tr>-->
+                        <!--<th></th>-->
+                        <!--<th>Actor-->
+                            <!--<p style="float:right;">-->
+                                <!--<input id="actorfilter-regex" size="12" type="text" placeholder="Filter..."/>-->
+                                <!--<a href="#" id="actorfilter-select-all">all</a> | <a href="#" id="actorfilter-select-none">none</a>-->
+                            <!--</p>-->
+                        <!--</th>-->
+                    <!--</tr>-->
+                    <!--</thead>-->
+                    <!--<tbody>-->
+                    <!--</tbody>-->
+                <!--</table>-->
+            <!--</div>-->
         </div>
         <div class="tab-pane fade panel panel-default" id="messagefiltering">
             <div class="panel-body" id="messagefilter">
